+++ src/pages/experiments/[experimentId]/plots-and-tables/dot-plot/index.jsx
@@ -22,7 +22,7 @@ import Header from 'components/Header';
 import PlotContainer from 'components/plots/PlotContainer';
 import Loader from 'components/Loader';
 import ExportAsCSV from 'components/plots/ExportAsCSV';
-import fileNames from 'utils/fileNames';
+import plotCsvFilename from 'utils/plotCsvFilename';
 import {
   updatePlotConfig,
   loadPlotConfig,
@@ -101,7 +101,7 @@ const DotPlotPage = (props) => {
   const highestGenesLoadedRef = useRef(false);
 
   const experimentName = useSelector((state) => state.experimentSettings.info.experimentName);
-  const csvFileName = fileNames(experimentName, 'DOT_PLOT', [config?.selectedCellSet, config?.selectedPoints]);
+  const csvFileName = plotCsvFilename(experimentName, 'DOT_PLOT', [config?.selectedCellSet, config?.selectedPoints]);
 
   useEffect(() => {
     if (!config) dispatch(loadPlotConfig(experimentId, plotUuid, plotType));
